requirements.txt
diskcache
gdown
tqdm
pandas
peewee
dtlpygeopandas
geopy
shapely
numpy
networkx
matplotlib
seaborn
rasterio
rapidfuzz
requests
statsmodels
scipy
urllib
google-cloud-storage
google-cloud-auth
pillow
mercantile
contextily
sshtunnel
scikit-learn

# Dev work, not directly used in the code:
pre-commit
black[d]
ipdb
